修改数据类型

+++ app/model/enterprise.go
@@ -68,21 +68,21 @@ type EnterpriseContact struct {
 
 // EnterpriseDetail 企业详细字段
 type EnterpriseDetail struct {
-	Name           *string    `json:"name" validate:"required" trans:"团签名称"`
-	CompanyName    *string    `json:"companyName" validate:"required" trans:"企业全称"`
-	Status         *uint8     `json:"status" enums:"0,1,2" validate:"required,min=0,max=2" trans:"合作状态"` // 0:未合作 1:合作中 2:已暂停
-	ContactName    *string    `json:"contactName" validate:"required" trans:"联系人"`
-	ContactPhone   *string    `json:"contactPhone" validate:"required" trans:"联系电话"`
-	IdcardNumber   *string    `json:"idcardNumber" validate:"required" trans:"身份证号"`
-	CityID         *uint64    `json:"cityId" validate:"required" trans:"所在城市"`
-	Address        *string    `json:"address" validate:"required" trans:"企业地址"`
-	Payment        *uint8     `json:"payment" validate:"required,min=1,max=2" enums:"1,2" trans:"付费方式"` // 1:预付费 2:后付费
-	Deposit        *float64   `json:"deposit" validate:"required" trans:"押金"`
-	Agent          *bool      `json:"agent"`              // 代理商 `true`是 `false`否
-	UseStore       *bool      `json:"useStore,omitempty"` // 可使用门店 `true`允许 `false`不允许 (仅代理商模式生效), 骑手是否可以使用门店进行激活和退租
-	Days           *[]int     `json:"days,omitempty"`     // 代理商时间选项
-	Distance       *uint64    `json:"distance"`           // 电柜距离
-	RechargeAmount *[]float64 `json:"rechargeAmount"`     // 充值金额选项
+	Name           *string  `json:"name" validate:"required" trans:"团签名称"`
+	CompanyName    *string  `json:"companyName" validate:"required" trans:"企业全称"`
+	Status         *uint8   `json:"status" enums:"0,1,2" validate:"required,min=0,max=2" trans:"合作状态"` // 0:未合作 1:合作中 2:已暂停
+	ContactName    *string  `json:"contactName" validate:"required" trans:"联系人"`
+	ContactPhone   *string  `json:"contactPhone" validate:"required" trans:"联系电话"`
+	IdcardNumber   *string  `json:"idcardNumber" validate:"required" trans:"身份证号"`
+	CityID         *uint64  `json:"cityId" validate:"required" trans:"所在城市"`
+	Address        *string  `json:"address" validate:"required" trans:"企业地址"`
+	Payment        *uint8   `json:"payment" validate:"required,min=1,max=2" enums:"1,2" trans:"付费方式"` // 1:预付费 2:后付费
+	Deposit        *float64 `json:"deposit" validate:"required" trans:"押金"`
+	Agent          *bool    `json:"agent"`              // 代理商 `true`是 `false`否
+	UseStore       *bool    `json:"useStore,omitempty"` // 可使用门店 `true`允许 `false`不允许 (仅代理商模式生效), 骑手是否可以使用门店进行激活和退租
+	Days           *[]int   `json:"days,omitempty"`     // 代理商时间选项
+	Distance       *uint64  `json:"distance"`           // 电柜距离
+	RechargeAmount *[]int   `json:"rechargeAmount"`     // 充值金额选项
 }
 
 type EnterpriseDetailWithID struct {
